Update index.html
I have added some explanation about endangered languages and degrees of endangerment of languages. Assignment 3 is not working, so I could see what is the result. I added the bibliography in the footer section. I am not sure it is a good idea.
Please feel free to modify the content.
Thank you.